+++ data/api_data.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class AccountsData(BaseModel):
+    user_name: str
+    password: str
+
